_assistant/backend/pyproject.toml
[project]
name = "hdx-assistant-backend"
version = "0.1.0"
description = "ClickHouse SQL Agent with LangGraph and AG-UI"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "langchain-chroma>=1.1.0",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langgraph-checkpoint-postgres>=3.0.2",
    "aiosqlite>=0.22.1",
    "langchain-classic>=1.0.1",
    "langchain-openai>=1.1.6",
    "langmem>=0.0.30",
    "langchain-mcp-adapters>=0.2.1",
    "psycopg[binary,pool]>=3.3.2",
    "aiosqlite>=0.22.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "clickhouse-connect>=0.7.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.25.0",
    "aiohttp>=3.13.2",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "fastmcp>=2.14.2",
#    "ag-ui-langgraph @ {root:parent:uri}/lib/ag_ui_langgraph",
    "ag-ui-protocol>=0.1.10",

]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hdx_agent", "src/ag_ui_langgraph"]

#[tool.uv.sources]
#ag-ui-langgraph = { path = "../lib/ag_ui_langgraph", editable = true }

[project.scripts]
backend = "hdx_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py311"
